BUGFIX: Multi-word pronunciations are correctly parsed
ISLEdict has multi-word pronunciations, which are generally
separated with underscores.  Before, only the first word was
analyzed with other words thrown out.  Now however, each
word is analyzed separately.  isleDict.lookup() now returns
a list with one element for each word in the multi-word search.

Note that to get a response with multiple words in it, one has to
explicitely search for an exact mutiple-word match.  E.g.
"how are you" is not in the dictionary and will not match but
"australian_seal" is in the dictionary and will match as two words.
Similarly, searching for "australian" will not match "australian_seal".