FEATURE: Naive word and phone alignment
Forced alignment is a useful feature.  Although it has to be
hand-corrected, it saves time by creating individual words and
phones in the textgrid.

But sometimes forced alignment might not be feasible or fails
to operate over some data.  Naive aligment can be used to create
the structure of the textgrid tiers for
words and phones that can then be easily adjusted in praat.

Under naive alignment all phones are set at a constant length
and each word is spaced